ML-API/requirements.txt
Flask==2.2.3
tensorflow==2.11.0
numpy==1.23.4
google-cloud-storage==2.10.0
google-auth==2.19.0
